+++ package.json
@@ -21,6 +21,7 @@
     "pg": "^8.11.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.22.3",
     "typescript": "5.0.4",
     "viem": "^1.21.4"
   },
